traveler/requirements.txt
Django==2.2.19
pytz==2024.2
sqlparse==0.5.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-phonenumber-field[phonenumberslite]